requirements.txt
asgiref==3.8.1
attrs==24.2.0
autopep8==1.6.0
Django==4.2.17
django-cors-headers==3.11.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
shortuuid==1.0.13
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.12.2
tzdata==2021.5
uritemplate==4.1.1
